Refactor BranchKeys and BroadcastIds.

+++ runtime/src/main/scala/com/asakusafw/spark/runtime/driver/BranchKey.scala
@@ -1,3 +1,3 @@
 package com.asakusafw.spark.runtime.driver
 
-case class BranchKey(id: Long)
+case class BranchKey(id: Int)
