+++ runtime/src/main/scala/com/asakusafw/spark/runtime/driver/BroadcastId.scala
@@ -1,3 +1,3 @@
 package com.asakusafw.spark.runtime.driver
 
-case class BroadcastId(id: Long)
+case class BroadcastId(id: Int)
